Working CAT control over USB host (both ways)

+++ software/T41_SDR/t41Control.h
@@ -13,4 +13,5 @@ extern uint8_t specData[518];
 void T41ControlSetup();
 void T41ControlLoop();
 void T41ControlSendData(uint8_t *data, int len);
-void SendSmeter(int16_t smeterPad, float32_t dbm);
+//void SendSmeter(int16_t smeterPad, float32_t dbm);
+void SendSmeter(int smeterPad, float dbm);
